requirements.txt
flask==2.3.3
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0
textract==1.6.3
python-magic==0.4.27
antiword==0.1.0
pytesseract==0.3.10
pdfminer.six==20181108 